CLN: refactored pyfibre.model.tools module
